还差我的订单和商品搜索，大功能 小功能还差 收货地址修改 我的信息修改

+++ src/main/java/com/app/shopping/mapper/ActivityMapper.java
@@ -19,7 +19,7 @@ public interface ActivityMapper {
      * @param id 主键
      * @return 实例对象
      */
-    Activity queryById(Long id);
+    Activity queryById(int id);
 
     /**
      * 查询指定行数据
